requirements.txt
discord.py==1.7.3
python-dotenv==0.18.0
pycodestyle==2.7.0